+++ packages/use-intl/src/core/TypesafeMessages.tsx
@@ -1,0 +1,30 @@
+export type NestedKeyOf<ObjectType> = ObjectType extends object
+  ? {
+      [Key in keyof ObjectType]:
+        | `${Key & string}`
+        | `${Key & string}.${NestedKeyOf<ObjectType[Key]>}`;
+    }[keyof ObjectType]
+  : never;
+
+export type NestedValueOf<
+  ObjectType,
+  Property extends string
+> = Property extends `${infer Key}.${infer Rest}`
+  ? Key extends keyof ObjectType
+    ? NestedValueOf<ObjectType[Key], Rest>
+    : never
+  : Property extends keyof ObjectType
+    ? ObjectType[Property]
+    : never;
+
+export type NamespaceKeys<ObjectType, Keys extends string> = {
+  [Property in Keys]: NestedValueOf<ObjectType, Property> extends string
+    ? never
+    : Property;
+}[Keys];
+
+export type MessageKeys<ObjectType, Keys extends string> = {
+  [Property in Keys]: NestedValueOf<ObjectType, Property> extends string
+    ? Property
+    : never;
+}[Keys];
