+++ packages/use-intl/src/core/createTranslatorImpl.tsx
@@ -1,9 +1,9 @@
 import AbstractIntlMessages from './AbstractIntlMessages.tsx';
 import {InitializedIntlConfig} from './IntlConfig.tsx';
+import {NestedKeyOf} from './TypesafeMessages.tsx';
 import createBaseTranslator from './createBaseTranslator.tsx';
 import {Formatters, IntlCache} from './formatters.tsx';
 import resolveNamespace from './resolveNamespace.tsx';
-import NestedKeyOf from './utils/NestedKeyOf.tsx';
 
 export type CreateTranslatorImplProps<Messages> = Omit<
   InitializedIntlConfig,
