+++ packages/use-intl/src/core/index.tsx
@@ -12,10 +12,12 @@ export {default as IntlError, IntlErrorCode} from './IntlError.tsx';
 export {default as createTranslator} from './createTranslator.tsx';
 export {default as createFormatter} from './createFormatter.tsx';
 export {default as initializeConfig} from './initializeConfig.tsx';
-export type {default as MessageKeys} from './utils/MessageKeys.tsx';
-export type {default as NamespaceKeys} from './utils/NamespaceKeys.tsx';
-export type {default as NestedKeyOf} from './utils/NestedKeyOf.tsx';
-export type {default as NestedValueOf} from './utils/NestedValueOf.tsx';
+export type {
+  MessageKeys,
+  NamespaceKeys,
+  NestedKeyOf,
+  NestedValueOf
+} from './TypesafeMessages.tsx';
 export {createIntlFormatters as _createIntlFormatters} from './formatters.tsx';
 export {createCache as _createCache} from './formatters.tsx';
 export type {default as AppConfig, Locale, Messages} from './AppConfig.tsx';
