--- packages/use-intl/src/core/utils/MessageKeys.tsx
@@ -1,9 +1,0 @@
-import NestedValueOf from './NestedValueOf.tsx';
-
-type MessageKeys<ObjectType, Keys extends string> = {
-  [Property in Keys]: NestedValueOf<ObjectType, Property> extends string
-    ? Property
-    : never;
-}[Keys];
-
-export default MessageKeys;
