--- packages/use-intl/src/core/utils/NestedKeyOf.tsx
@@ -1,9 +1,0 @@
-type NestedKeyOf<ObjectType> = ObjectType extends object
-  ? {
-      [Key in keyof ObjectType]:
-        | `${Key & string}`
-        | `${Key & string}.${NestedKeyOf<ObjectType[Key]>}`;
-    }[keyof ObjectType]
-  : never;
-
-export default NestedKeyOf;
