--- packages/use-intl/src/core/utils/NestedValueOf.tsx
@@ -1,12 +1,0 @@
-type NestedValueOf<
-  ObjectType,
-  Property extends string
-> = Property extends `${infer Key}.${infer Rest}`
-  ? Key extends keyof ObjectType
-    ? NestedValueOf<ObjectType[Key], Rest>
-    : never
-  : Property extends keyof ObjectType
-    ? ObjectType[Property]
-    : never;
-
-export default NestedValueOf;
