+++ packages/use-intl/src/react/useTranslationsImpl.tsx
@@ -1,8 +1,8 @@
 import {useMemo} from 'react';
 import AbstractIntlMessages from '../core/AbstractIntlMessages.tsx';
+import {NestedKeyOf} from '../core/TypesafeMessages.tsx';
 import createBaseTranslator from '../core/createBaseTranslator.tsx';
 import resolveNamespace from '../core/resolveNamespace.tsx';
-import NestedKeyOf from '../core/utils/NestedKeyOf.tsx';
 import {IntlError, IntlErrorCode} from '../core.tsx';
 import useIntlContext from './useIntlContext.tsx';
